Crash when we should crash
We should only use the ? operator for recoverable errors. It
effectively results in the error being rethrown up the stack. It makes
debugging hard since the original stacktrace won't be available when a
result is finally unwrapped. The stacktrace displayed at that point will
start at the place where it was unwrapped.